+++ tests/conftest.py
@@ -246,26 +246,6 @@ def fixture_parameters(
             name="Conti_finite2",
             bounds=(-1, 0),
         ),
-        NumericContinuous(
-            name="Conti_infinite1",
-            bounds=(None, 1),
-        ),
-        NumericContinuous(
-            name="Conti_infinite2",
-            bounds=(0, None),
-        ),
-        NumericContinuous(
-            name="Conti_infinite3",
-            bounds=(0, np.inf),
-        ),
-        NumericContinuous(
-            name="Conti_infinite4",
-            bounds=(-np.inf, 1),
-        ),
-        NumericContinuous(
-            name="Conti_infinite5",
-            bounds=(None, None),
-        ),
         Custom(
             name="Custom_1",
             data=pd.DataFrame(